+++ src/packagedcode/models.py
@@ -1043,9 +1043,9 @@ class DatafileHandler:
     # possible values are app, sys and info
     datasource_type = 'app'
 
-    # tuple of specifically supported operating systems. If None or empty, all platforms are supported
-    # possible values are win, mac, linux, freebsd
-    supported_oses = tuple()
+    # tuple of operating systems where getting package data using this DatafileHandler is supported.
+    # If None or empty, all platforms are supported, possible values are win, mac, linux, freebsd
+    supported_oses = None
 
     # Sequence of known fnmatch-style case-insensitive glob patterns (e.g., Unix
     # shell style patterns) that apply on the whole POSIX path for package