+++ tests/scancode/test_cli.py
@@ -128,7 +128,7 @@ def test_no_version_check_run_is_successful():
 
 
 def test_usage_and_help_return_a_correct_script_name_on_all_platforms():
-    result = run_scan_click(['--help'])
+    result = run_scan_click(options=['--help'], test_mode=False, retry=False)
     assert 'Usage: scancode [OPTIONS]' in result.output
     # this was showing up on Windows
     assert 'scancode-script.py' not in result.output